feat: implement minimal stack navigation
