+++ examples/objectdetection/package.json
@@ -13,9 +13,13 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.9.26",
     "@shopify/react-native-skia": "^1.0.5",
     "react": "18.2.0",
     "react-native": "0.73.6",
+    "react-native-safe-area-context": "^4.9.0",
+    "react-native-screens": "^3.30.1",
     "react-native-vision-camera": "^3.9.2",
     "react-native-worklets-core": "^0.4.0"
   },
