+++ examples/objectdetection/src/StillPhoto.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+export default function StillPhoto(): React.ReactElement | null {
+  return (
+    <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
+      <Text>Still Photo Placeholder</Text>
+    </View>
+  );
+}
